+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `FormerlySerializedAsBackingFieldOfAttribute` that correctly formats the name to match the format `<{propertyName}>k__BackingField`.
 - Added `DebuggerStepThroughAttribute` where applicable.
 - Added `EditorBrowsableAttribute(EditorBrowsableState.Never)` where applicable.
+- Changed static `ManagedPool` to require the `IDisposable` interface indirectly through the `ISharedManagedPoolHandler` interface.
 - Fixed `Disposable<T>` allowing multiple `Dispose` calls.
 
 ## [10.2.10] - 2022-05-24